Resize on more parts of the vertical line
