+++ ui/src/frontend/notes_panel.ts
@@ -38,6 +38,7 @@ import {
 import {Panel, PanelSize} from './panel';
 import {isTraceLoaded} from './sidebar';
 import {customButtonRegistry} from './button_registry';
+import {resizeTrackShell} from './vertical_line_helper';
 
 const FLAG_WIDTH = 16;
 const AREA_TRIANGLE_WIDTH = 10;
@@ -74,35 +75,6 @@ export class NotesPanel extends Panel {
     }, {passive: true});
   }
 
-  resize = (e: MouseEvent): void => {
-    e.stopPropagation();
-    e.preventDefault();
-    const mouseMoveEvent = (evMove: MouseEvent): void => {
-        evMove.preventDefault();
-        const root = document.querySelector(':root');
-        if (root && root instanceof HTMLElement &&
-            'layerX' in evMove && evMove.layerX &&
-            typeof evMove.layerX === 'number'
-        ) {
-          if (evMove.layerX < 250) {
-            root.style.setProperty('--track-shell-width', '250px');
-          } else {
-            root.style.setProperty('--track-shell-width', evMove.layerX + 'px');
-          }
-          globals.rafScheduler.scheduleFullRedraw();
-        }
-    };
-    const mouseUpEvent = (evUp : MouseEvent): void => {
-        evUp.stopPropagation();
-        evUp.preventDefault();
-        document.removeEventListener('mousemove', mouseMoveEvent);
-        document.removeEventListener('mouseup', mouseUpEvent);
-    };
-    document.addEventListener('mousemove', mouseMoveEvent);
-    document.addEventListener('mouseup', mouseUpEvent);
-    document.removeEventListener('mousedown', this.resize);
-  };
-
   view() {
     const allCollapsed = Object.values(globals.state.trackGroups)
                              .every((group) => group.collapsed);
@@ -121,18 +93,18 @@ export class NotesPanel extends Panel {
                 (e.layerX -2) <= (getCssNum('--track-shell-width') || 0)
               )
             ) {
-              document.addEventListener('mousedown', this.resize);
+              document.addEventListener('mousedown', resizeTrackShell);
               e.currentTarget.style.cursor = 'col-resize';
               return;
             } else if (e.currentTarget instanceof HTMLElement) {
               e.currentTarget.style.cursor = 'unset';
             }
-            document.removeEventListener('mousedown', this.resize);
+            document.removeEventListener('mousedown', resizeTrackShell);
           },
           onmouseleave: (e: PerfettoMouseEvent) =>{
             if (e.currentTarget instanceof HTMLElement) {
               e.currentTarget.style.cursor = 'unset';
-              document.removeEventListener('mousedown', this.resize);
+              document.removeEventListener('mousedown', resizeTrackShell);
             }
           },
           oncontextmenu: (e: PerfettoMouseEvent)=>{
