+++ ui/src/frontend/time_selection_panel.ts
@@ -33,6 +33,8 @@ import {
   timeScaleForVisibleWindow,
 } from './gridline_helper';
 import {Panel, PanelSize} from './panel';
+import {PerfettoMouseEvent} from './events';
+import {resizeTrackShell} from './vertical_line_helper';
 
 export interface BBox {
   x: number;
@@ -128,7 +130,30 @@ function drawIBar(
 
 export class TimeSelectionPanel extends Panel {
   view() {
-    return m('.time-selection-panel');
+    return m('.time-selection-panel', {
+
+      onmousemove: (e: PerfettoMouseEvent)=>{
+        if (e.currentTarget instanceof HTMLElement &&
+          (
+            (e.layerX +2) >= (getCssNum('--track-shell-width') || 0) &&
+            (e.layerX -2) <= (getCssNum('--track-shell-width') || 0)
+          )
+        ) {
+          document.addEventListener('mousedown', resizeTrackShell);
+          e.currentTarget.style.cursor = 'col-resize';
+          return;
+        } else if (e.currentTarget instanceof HTMLElement) {
+          e.currentTarget.style.cursor = 'unset';
+        }
+        document.removeEventListener('mousedown', resizeTrackShell);
+      },
+      onmouseleave: (e: PerfettoMouseEvent) =>{
+        if (e.currentTarget instanceof HTMLElement) {
+          e.currentTarget.style.cursor = 'unset';
+          document.removeEventListener('mousedown', resizeTrackShell);
+        }
+      },
+    });
   }
 
   renderCanvas(ctx: CanvasRenderingContext2D, size: PanelSize) {
